requirements.txt
plotly=4.13.0
scipy=1.5.2
streamlit-nightly==0.82.1.dev20210527

